// packages/service-payment-mock-server/src/payment-interface.ts
export interface CustomerProfile {
  customerId?: string;
  billingFirstName: string;
  billingLastName: string;
  billingEmail: string;
  billingPhone?: string;
  billingAddressLine1: string;
  billingAddressLine2?: string;
  billingCity: string;
  billingState: string;
  billingPostalCode: string;
  billingCountry: string;
}

export interface PaymentTokenInfo {
  paymentToken: string;
  isDefault: boolean;
}

export interface PaymentInstrumentInfo {
  id: string;
  paymentInstrumentId: string;
  cardType: string;
  cardExpirationMonth: string;
  cardExpirationYear: string;
}

export interface PaymentTransactionResponse {
  _links: {
    self: {
      href: string;
      method?: string;
    };
    capture: {
      href: string;
      method?: string;
    };
  };
  id: string;
  submitTimeUtc: string; // '2024-07-08T20:56:04Z'
  status: string; // 'AUTHORIZED'
  reason: string; // 'MISSING_FIELD' | 'INVALID_DATA' | 'INVALID_CARD'
  reconciliationId: string;
  clientReferenceInformation: { code: string };
  processorInformation: {
    approvalCode: string;
    responseCode: string;
    avs: { code: string; codeRaw: string };
    cardVerification: { resultCode: string };
    paymentAccountReferenceNumber: string;
    transactionId: string;
    networkTransactionId: string;
  };
  paymentAccountInformation: { card: { type: string } }; // type: '001', '002', etc.
  paymentInformation: {
    card: { type: string };
    tokenizedCard: { type: string };
    paymentInstrument: { id: string };
    instrumentIdentifier: { id: string; state: string }; // state: 'ACTIVE'
    accountFeatures: { balanceAmount: string };
    customer: { id: string };
  };
  orderInformation: {
    amountDetails: {
      totalAmount: string;
      authorizedAmount: string;
      currency: string;
    };
  };
  pointOfSaleInformation: { terminalId: string };
  tokenInformation: {
    instrumentidentifierNew: boolean;
    paymentInstrument: { id: string };
    instrumentIdentifier: { id: string; state: string }; // state: 'ACTIVE'
    customer: { id: string };
  };
  voidAmountDetails: { voidAmount: string; currency: string };
}

export interface TransactionReceipt {
  isSuccess: boolean;
  vendor?: string;
  completedAt?: Date;
  amount?: number;
  transactionId?: string;
  reconciliationId?: string;
  errorOccurredAt?: Date;
  errorCode?: string;
  errorMessage?: string;
}

export interface PaymentInstrument {
  _links: {
    self: {
      href: string;
    };
    customer: {
      href: string;
    };
  };
  id: string;
  object: string; // 'paymentInstrument'
  default: boolean;
  state: string; // 'ACTIVE'
  card: {
    expirationMonth: string;
    expirationYear: string;
    type: string; // '001' = 'Visa', '002' = 'Master Card', '003' = 'American Express'
  };
  buyerInformation: {
    currency: string;
  };
  billTo: {
    firstName: string;
    lastName: string;
    address1: string;
    address2: string;
    locality: string;
    administrativeArea: string;
    postalCode: string;
    country: string;
    email: string;
    phoneNumber: string;
  };
  processingInformation: {
    billPaymentProgramEnabled: boolean;
  };
  instrumentIdentifier: {
    id: string;
  };
  metadata: {
    creator: string;
  };
  _embedded: {
    instrumentIdentifier: {
      _links: {
        self: {
          href: string;
        };
        paymentInstruments: {
          href: string;
        };
      };
      id: string;
      object: string; // 'instrumentIdentifier'
      state: string; // 'ACTIVE'
      card: {
        number: string;
      }; // '222242XXXXXX1113'
      processingInformation: {
        authorizationOptions: {
          initiator: {
            merchantInitiatedTransaction: {
              previousTransactionId: string;
            };
          };
        };
      };
      metadata: {
        creator: string; // 'ecfmg_faimer'
      };
    };
  };
}

export interface CustomerPaymentResponse {
  _links: {
    self: {
      href: string;
    };
    paymentInstruments: {
      href: string;
    };
  };
  id: string; // Customer ID
  clientReferenceInformation: { code: string };
  defaultPaymentInstrument: { id: string };
  metadata: { creator: string };
  _embedded: {
    defaultPaymentInstrument: PaymentInstrument;
  };
}

export interface CustomerPaymentInstrumentResponse extends PaymentInstrument {}

export interface CustomerPaymentInstrumentsResponse {
  _links: {
    self: {
      href: string;
    };
    first: {
      href: string;
    };
    last: {
      href: string;
    };
  };
  offset: number;
  limit: number;
  count: number;
  total: number;
  _embedded: {
    paymentInstruments: PaymentInstrument[];
  };
}